pyproject.toml
[project]
name = "mcp-google-tasks"
version = "0.1.0"
description = "FastMCP server exposing Google Tasks as tools."
requires-python = ">=3.10"
dependencies = [
  "fastmcp>=0.3.0",
  "google-api-python-client>=2.153.0",
  "google-auth>=2.35.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.1",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.1",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
fastmcp = { git = "https://github.com/jlowin/fastmcp.git" }

